requirements.txt
Django==1.5.4
PIL==1.1.7
Pillow==2.2.2
argparse==1.2.1
django-crispy-forms==1.4.0
django-debug-toolbar==0.11.0
django-notify==1.1.2
django-pagination==1.0.7
django-selectable==0.8.0
django-taggit==0.11.2
ipdb==0.8
ipython==0.13.2
psycopg2==2.4.5
python-dateutil==2.2
six==1.6.1
sqlparse==0.1.10
wsgiref==0.1.2
